requirements.txt
# ============================================================
# Party Master Data Hub — Python Dependencies
# ============================================================

# --- Web Framework & API ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# --- Database ---
psycopg2-binary==2.9.9          # PostgreSQL driver
sqlalchemy==2.0.25               # ORM / query builder
alembic==1.13.1                  # DB migrations

# --- AWS / LocalStack Client ---
boto3==1.34.14                   # AWS SDK (works with LocalStack too)
botocore==1.34.14

# --- Kafka / Streaming ---
confluent-kafka==2.3.0           # Kafka producer/consumer

# --- Data Processing ---
pyspark==3.5.0                   # Local Glue simulation
pandas==2.1.4
pyarrow==14.0.2                  # Parquet file support
numpy==1.26.3

# --- Data Mastering / Identity Resolution ---
recordlinkage==0.16              # Fuzzy matching and deduplication
# dedupe==3.0.3                  # Skipped - requires Microsoft C++ Build Tools on Windows
                                 # recordlinkage above covers the same use case for our project
jellyfish==1.0.3                 # String similarity (Jaro-Winkler, Soundex)
fuzzywuzzy==0.18.0               # Fuzzy string matching
python-Levenshtein==0.23.0       # Fast Levenshtein distance

# --- Data Quality ---
great-expectations==0.18.8       # Data validation framework

# --- Data Generation (Fake Party Data) ---
faker==22.0.0                    # Generate realistic fake names, addresses, companies

# --- Utilities ---
python-dotenv==1.0.0             # Load .env config files
loguru==0.7.2                    # Better logging
httpx==0.26.0                    # Async HTTP client (for API mocking)
requests==2.31.0                 # Sync HTTP client
tenacity==8.2.3                  # Retry logic for resilient pipelines
schedule==1.2.1                  # Simple job scheduling

# --- Testing ---
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0